Version Two
In this version I focused on getting the input for the user to order a pizza. The user will enter the number that corresponds to the pizza that they want to order, and the program will store this and output the price of the pizza and the id that they inputted. At the moment, it only takes in the order for one pizza, but that will change later.