Version Three
In this version I added the capability for multiple pizzas to be ordered at once, and then calculated the total price of the order. The user enters the number of pizzas that they want to order, and each time the menus print out and the user enters what pizza they want.