Version Five
In this version I added the option for users to add stuffed crust to any of their pizzas. After every pizza, they are prompted as to whether or not they want to add stuffed crust to that pizza for the price of 3 dollars. If they answer yes, this price is added to the price variable. Regardless of what they answer, their answer gets appended to a list that keeps track of whether or not each pizza has stuffed crust or not. This is included in the output of  the user’s order.