Version Six
In this version I added the option of either delivery or pickup for the order. I put this in a function that gets called at the very beginning of the program. If the user enters D for delivery an extra eight dollars gets added onto the total price. Otherwise, if they enter P, the price is not affected.