Version Seven
In this version I added the name input and a confirmation process at the end of the program. The name input is done at the start of the program (and if the user edits their delivery details later), and the confirmation process allows the user to change their pizza order and delivery details before they submit the order.