Version Eight
In this version I added error handling for the two sets boundaries in the program – the amount of pizzas that the program is ordering, and the ID of the pizzas. If they try to order more than 3 pizzas the program will stop them and say that it is due to the pandemic. On the other end, if they try to order no pizzas the program will tell them that they have to order at least one pizza. If they try to order a pizza that doesn’t exist the program will ask them to enter a valid pizza ID and send them back to the menu input.